# backend/src/doc/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
			"ISomeConstant": {
				"type": "string",
				"enum": [
					"userRole",
					"gender",
					"anything unchangeable"
				]
			},
			"IBoilerplateCode": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"constant": {
						"$ref": "#/components/schemas/ISomeConstant"
					},
					"removeIt": {
						"type": "boolean"
					},
					"anyRef": {
						"$ref": "#/components/schemas/Types.ObjectId"
					},
					"_id": {
						"$ref": "#/components/schemas/Types.ObjectId"
					}
				},
				"required": [
					"constant",
					"removeIt",
					"anyRef"
				],
				"type": "object"
			},
			"IApiResponse_IBoilerplateCode_": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IBoilerplateCode"
							}
						],
						"nullable": true
					},
					"meta": {
						"properties": {
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"total",
							"limit",
							"page"
						],
						"type": "object",
						"nullable": true
					},
					"message": {
						"type": "string",
						"nullable": true
					},
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"statusCode"
				],
				"type": "object"
			},
			"IApiResponse_null_": {
				"properties": {
					"data": {
						"anyOf": [],
						"nullable": true
					},
					"meta": {
						"properties": {
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"total",
							"limit",
							"page"
						],
						"type": "object",
						"nullable": true
					},
					"message": {
						"type": "string",
						"nullable": true
					},
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"statusCode"
				],
				"type": "object"
			},
			"Partial_IBoilerplateCode_": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/Types.ObjectId"
					},
					"anyRef": {
						"$ref": "#/components/schemas/Types.ObjectId"
					},
					"removeIt": {
						"type": "boolean"
					},
					"constant": {
						"$ref": "#/components/schemas/ISomeConstant"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "English Sphere",
		"version": "1.0.0",
		"description": "All api documentation",
		"license": {
			"name": "License: Apache 2.0"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/boilerplateCode/{userId}": {
			"get": {
				"operationId": "GetBoilerplateCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiResponse_IBoilerplateCode_"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing user.",
				"tags": [
					"BoilerplateCodeController"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "X-Access-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Supply the valid user mongodb ID.",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/boilerplateCode": {
			"post": {
				"operationId": "AddBoilerplateCode",
				"responses": {
					"201": {
						"description": "created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Adds a new user by providing required request body.",
				"tags": [
					"BoilerplateCodeController"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "X-Access-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "_id, createdAt and updated are optional fields",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBoilerplateCode",
								"description": "_id, createdAt and updated are optional fields"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateBoilerplateCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Patch docs here",
				"tags": [
					"BoilerplateCodeController"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "X-Access-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IBoilerplateCode_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveBoilerplateCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Delete docs here",
				"tags": [
					"BoilerplateCodeController"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "X-Access-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"value": {
										"type": "string"
									}
								},
								"required": [
									"value"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}